feat(core): add scheduled service base
